chore(suite-desktop): move discovery selectors
