+++ packages/suite/src/components/suite/modals/Passphrase/index.tsx
@@ -1,16 +1,18 @@
-import TrezorConnect from '@trezor/connect';
 import React, { useCallback, useState } from 'react';
+import { useIntl } from 'react-intl';
+
 import styled from 'styled-components';
-import { variables, PassphraseTypeCard } from '@trezor/components';
 import { useSelector, useDispatch } from 'src/hooks/suite';
 import { onPassphraseSubmit } from 'src/actions/suite/modalActions';
-import { selectIsDiscoveryAuthConfirmationRequired } from 'src/reducers/wallet/discoveryReducer';
+import { selectIsDiscoveryAuthConfirmationRequired } from 'src/reducers/suite/suiteReducer';
 import * as deviceUtils from 'src/utils/suite/device';
 import { Translation, Modal } from 'src/components/suite';
 import type { TrezorDevice } from 'src/types/suite';
 import { OpenGuideFromTooltip } from 'src/components/guide';
-import { useIntl } from 'react-intl';
 import messages from 'src/support/messages';
+
+import { variables, PassphraseTypeCard } from '@trezor/components';
+import TrezorConnect from '@trezor/connect';
 
 const Wrapper = styled.div<{ authConfirmation?: boolean }>`
     display: flex;
